refactor(desktop): 拆出 mcps 设置面板
将 McpsSettingsPanel 及 MCP 辅助函数迁入 settings/panels/mcps-settings-panel.tsx。
